WEB-678: Restructure Reporting Plugin to Support Branch Model Operations
