+++ src/app/loans/common-resolvers/loans-transaction-reciept.resolver.ts
@@ -14,7 +14,7 @@ import { ActivatedRouteSnapshot } from '@angular/router';
 import { Observable } from 'rxjs';
 
 /** Custom Services */
-import { ReportsService } from 'app/reports/reports.service';
+import { ReportsService } from 'app/reporting-plugin/services/reports.service';
 
 /**
  * Loans Transaction Reciept resolver.
