+++ src/app/reporting-plugin/resolvers/reports.resolver.ts
@@ -6,6 +6,14 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
+/**
+ * Copyright since 2025 Mifos Initiative
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+
 /** Angular Imports */
 import { Injectable, inject } from '@angular/core';
 
@@ -13,7 +21,7 @@ import { Injectable, inject } from '@angular/core';
 import { Observable } from 'rxjs';
 
 /** Custom Services */
-import { ReportsService } from '../reports.service';
+import { ReportsService } from '../services/reports.service';
 
 /**
  * Reports data resolver.
