+++ src/app/reports/run-report/birt/birt.component.ts
@@ -15,7 +15,7 @@ import { finalize, catchError } from 'rxjs/operators';
 import { of } from 'rxjs';
 
 /** Custom Services */
-import { ReportsService } from '../../reports.service';
+import { ReportsService } from 'app/reporting-plugin/services/reports.service';
 import { SettingsService } from 'app/settings/settings.service';
 import { ProgressBarService } from 'app/core/progress-bar/progress-bar.service';
 import { STANDALONE_SHARED_IMPORTS } from 'app/standalone-shared.module';
