+++ src/app/savings/savings-account-view/transactions-tab/export-transactions/export-transactions.component.ts
@@ -13,7 +13,7 @@ import { UntypedFormBuilder, Validators, ReactiveFormsModule } from '@angular/fo
 import { ActivatedRoute, RouterLink } from '@angular/router';
 
 /** Custom Services */
-import { ReportsService } from 'app/reports/reports.service';
+import { ReportsService } from 'app/reporting-plugin/services/reports.service';
 import { SettingsService } from 'app/settings/settings.service';
 import { Dates } from 'app/core/utils/dates';
 import { FaIconComponent } from '@fortawesome/angular-fontawesome';
